dev/test/requirements.txt
# Test-only dependencies
httpx>=0.27.0
pytest>=8.0.0
pytest-timeout>=2.3.0
pytest-xdist>=3.5.0
python-dotenv>=1.0.1
